Manual

Introduction

The core idea behind this package is to provide a toolbox for developing and integrating QUBO annealing/sampling tools with the JuMP mathematical programming environment. Appart from the few couple bundled engines, Anneal.jl is inherently about extensions, which is achieved by implementing most of the MOI requirements, leaving only the essential for the developer.

The annealers and samplers defined via the AbstractSampler{T} interface only support models given in a QUBO form, as explained below.

QUBO

An optimization problem is in its QUBO form if it is written as

\[\begin{array}{rl} \min & \mathbf{x}^{\intercal} \mathbf{Q}\, \mathbf{x} + c \\ \text{s.t.} & \mathbf{x} \in \mathbb{B}^{n} \end{array}\]

where $\mathbf{Q} \in \mathbb{R}^{n}$ is a symmetric matrix and $c \in \mathbb{R}$ the constant offset.

In terms of Julia's data structures, we define a QUBO Normal Form (NQF) with a triplet

(x::Dict{VI, Union{Int, Nothing}}, Q::Dict{Tuple{Int, Int}, T}, c::T)

where $x$ provides a mapping between each MathOptInterface's VariableIndex and the corresponding integer index in Q.

Even though only QUBO formulations are supported as input, Anneal provides internal tools for working with Ising Model instances since many samplers rely on it. Model validation and trivial QUBO/Insing conversion is made using the functions below.

Anneal.isqubo — Function

isqubo(model::MOI.ModelLike)

Tells if a given model is ready to be interpreted as a QUBO model.

For it to be true, a few conditions must be met:

All variables must be binary (MOI.VariableIndex ∈ MOI.ZeroOne)

No other constraints are allowed

The objective function must be of type MOI.ScalarQuadraticFunction, MOI.ScalarAffineFunction or MOI.VariableIndex

The objective sense must be either MOI.MIN_SENSE or MOI.MAX_SENSE

Anneal.qubo_normal_form — Function

qubo_normal_form(model::MOI.ModelLike)
qubo_normal_form(T::Type, model::MOI.ModelLike)

Returns a triple $(x, Q, c)$ where:

x::Dict{MOI.VariableIndex, Union{Int, Nothing}} maps each of the model's variables to an integer index, to be used when interacting with Q.

Q::Dict{Tuple{Int, Int}, T} is a sparse representation of the QUBO Matrix.

c::T is the constant term associated with the problem.

Anneal.ising_normal_form — Function

ising_normal_form(model::MOI.ModelLike)
ising_normal_form(T::Type, model::MOI.ModelLike)
ising_normal_form(x::Dict{VI, Union{Int, Nothing}}, Q::Dict{Tuple{Int, Int}, T}, c::T) where {T}

Returns a quadruple $(s, h, J, c)$ where:

s::Dict{MOI.VariableIndex, Union{Int, Nothing}} maps each of the model's variables to an integer index, to be used when interacting with $h$ and $J$.

h::Dict{Int, T} is a sparse vector for the linear terms of the Ising Model.

J::Dict{Tuple{Int, Int}, T} is a sparse representation of the quadratic magnetic interactions.

c::T is the constant term associated with the problem.

Defining a new sampler interface

The Anneal.@anew macro is available to speed up the interface setup process.

Anneal.@anew — Macro

    @anew(expr)

The `@anew` macro receives a `begin ... end` block with an attribute definition on each of the block's statements.

All attributes must be presented as an assignment to the default value of that attribute. To create a MathOptInterface optimizer attribute, an identifier must be present on the left hand side. If a solver-specific, raw attribute is desired, its name must be given as a string, e.g. between double quotes. In the special case where an attribute could be accessed in both ways, the identifier must be followed by the parenthesised raw attribute string. In any case, the attribute type can be specified typing the type assertion operator `::` followed by the type itself just before the equal sign.

For example, a list of the valid syntax variations for the *number of reads* attribute follows:
    - "num_reads" = 1_000
    - "num_reads"::Integer = 1_000
    - NumberOfReads = 1_000
    - NumberOfReads::Integer = 1_000
    - NumberOfReads("num_reads") = 1_000
    - NumberOfReads("num_reads")::Integer = 1_000

Inside a module scope for the new interface, one should call the Anneal.@anew macro, specifying the solver's attributes as described in the macro's docs. One must also define MOI.get methods for the MOI.SolverName, MOI.RawSolver and MOI.SolverVersion attributes. The last and most important step is to define the Anneal.sample method, which returns both a vector with every sample and also the sampling time. The whole standard definition is described in the next example.

module SuperAnnealer
    using Anneal

    Anneal.@anew begin
        NumberOfReads::Integer = 100
        SuperAttribute::Any = nothing
    end # This will define Optimizer{T} <: MOI.AbstractOptimizer

    # -*- MathOptInterface -*-
    function MOI.get(::Optimizer, MOI.SolverName)
        "Super Annealer"
    end

    function MOI.get(::Optimizer, MOI.SolverVersion)
        v"0.1.0" # Version strings are welcome!
    end

    function MOI.get(sampler::Optimizer, MOI.RawSolver)
        # Usually the sampler itself, since we are prtty much low-level here.
        # If you call another optimizer under the hood, you might want to return it.
        sampler 
    end

    function Anneal.sample(sampler::Optimizer)
        # Is your annealer running on the Ising Model? Have this:
        s, h, J, c = Anneal.ising_normal_form(sampler.x, sampler.Q, sampler.c)

        n = MOI.get(sampler, NumberOfReads())
        attr = MOI.get(sampler, SuperAttribute())

        t0 = time()
        results = [super_sample(h, J, attr) for i = 1:n]
        t1 = time()

        return (results, t1 - t0)
    end
end

Type assertion for Anneal.sample's return can be done using the ::Tuple{Anneal.SamplerResults, Float64} query.