Samplers

Abstract Sampler

Anneal.AbstractSampler — Type

AbstractSampler{T} <: MOI.AbstractOptimizer

Identity Sampler

Anneal.IdentitySampler.Optimizer — Type

IdentitySampler.Optimizer{T}

This sampler selects precisely the state vector provided as warm-start.

Exact Sampler

Anneal.ExactSampler.Optimizer — Type

ExactSampler.Optimizer{T}

This sampler performs an exhaustive search over all $2^{n}$ possible states.

Warn

Due to the exponetially large amount of visited states, it is not possible to use this sampler for problems any larger than $20$ variables big.

Random Sampler

Anneal.RandomSampler.Optimizer — Type

RandomSampler.Optimizer{T}

Attributes

RandomSeed, "seed": Random seed to initialize the random number generator.

NumberOfReads, "num_reads": Number of random states sampled per run.

RandomGenerator, "rng": Random Number Generator instance.